requirements.txt
asgiref==3.8.1
Django==5.0.4
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-restframework==0.0.1
djangorestframework==3.15.1
mysqlclient==2.2.4
pillow==10.3.0
python-decouple==3.8
sqlparse==0.5.0
tzdata==2024.1
